Convert Request to Ajax Request

+++ database/migrations/2021_02_11_123304_create_bookings_table.php
@@ -19,7 +19,7 @@ class CreateBookingsTable extends Migration
             $table->date('arrival_date');
             $table->date('departure_date');
             $table->integer('persons_num');
-            $table->float('total_cost')->default('0');
+            $table->double('total_cost')->default('0');
             $table->enum('status',
                 ['pending','completed','rejected'])->default('pending');
             $table->integer('booking_option_id');
